refactor: добавлен индексный файл в utils для реэкспорта
