+++ packages/client/src/pages/gamePage/utils/roadside/index.ts
@@ -1,0 +1,1 @@
+export { Tree } from './Tree';
